+++ R/eem_utils.R
@@ -296,7 +296,8 @@ eem_set_wavelengths <- function(eem, ex, em){
 #' @param sample Either numeric of character vector. See \code{details} for more
 #'   information.
 #'
-#' @param remove logical. Should EEMs removed (TRUE) or extracted (FALSE).
+#' @param keep logical. If TRUE, the specified \code{sample} will be returned.
+#'   If FALSE, they will be removed.
 #'
 #' @param ignore_case Logical, should sample name case should be ignored (TRUE)
 #'   or not (FALSE). Default is FALSE.
@@ -315,8 +316,13 @@ eem_set_wavelengths <- function(eem, ex, em){
 #' folder <- system.file("extdata/cary/scans_day_1", package = "eemR")
 #' eems <- eem_read(folder)
 #'
-#' eem_extract(eems, c(1, 3)) ## Removes samples 1 and 3
-#' eem_extract(eems, c(1, 3), remove = TRUE) ## extract samples 1 and 3
+#' eems
+#'
+#' # Remove first and third samples
+#' eem_extract(eems, c(1, 3))
+#'
+#' # Remove everything except first and third samples
+#' eem_extract(eems, c(1, 3), keep = TRUE)
 #'
 #' # Remove all samples containing "3" in their names.
 #' eem_extract(eems, "3")
@@ -327,11 +333,8 @@ eem_set_wavelengths <- function(eem, ex, em){
 #' # Remove all samples containing "blank" or "nano"
 #' eem_extract(eems, c("blank", "nano"))
 #'
-#' # Remove all samples starting with "no"
-#' eem_extract(eems, "^s")
-#'
 #' @export
-eem_extract <- function(eem, sample, remove = FALSE, ignore_case = FALSE,
+eem_extract <- function(eem, sample, keep = FALSE, ignore_case = FALSE,
                         verbose = TRUE) {
 
   stopifnot(class(eem) == "eemlist",
@@ -344,10 +347,14 @@ eem_extract <- function(eem, sample, remove = FALSE, ignore_case = FALSE,
 
     stopifnot(all(is_between(sample, 1, length(eem))))
 
-    eem[ifelse(remove, -sample, sample)] <- NULL
+    to_remove <- ifelse(rep(keep, length(sample)),
+                        setdiff(1:length(eem), sample),
+                        sample)
+
+    eem[to_remove] <- NULL
 
     if(verbose){
-      cat(ifelse(remove, "Removed sample(s):", "Extracted sample(s):"),
+      cat(ifelse(keep, "Extracted sample(s):", "Removed sample(s):"),
           sample_names[sample], "\n")
     }
   }
@@ -355,19 +362,19 @@ eem_extract <- function(eem, sample, remove = FALSE, ignore_case = FALSE,
   ## Regular expression
   if(is.character(sample)){
 
-    index <- grepl(paste(sample, collapse = "|"),
+    to_remove <- grepl(paste(sample, collapse = "|"),
                    sample_names,
                    ignore.case = ignore_case)
 
-    eem[xor(index, !remove)] <- NULL
+    eem[xor(to_remove, keep)] <- NULL
 
     if(verbose){
-      if(all(index == FALSE)){
+      if(all(to_remove == FALSE)){
         cat("Nothing to remove.")
       }
       else{
-        cat(ifelse(remove, "Removed sample(s):", "Extracted sample(s):"),
-            sample_names[index], "\n")
+        cat(ifelse(keep, "Extracted sample(s):", "Removed sample(s):"),
+            sample_names[to_remove], "\n")
       }
     }
   }
@@ -610,7 +617,7 @@ eem_extract_blank <- function(eem, average = TRUE) {
   blank_names <- c("nano", "miliq", "milliq", "mq", "blank")
 
   blank <- eem_extract(eem, blank_names,
-                       remove = FALSE,
+                       keep = TRUE,
                        ignore_case = TRUE,
                        verbose = FALSE)